==> 02_Analysis_Overall.R <==
source("utils.R")

## Creating list of packages to load
package_list <- c("elliottmess/butteR", "hypegrammaR","ellieallien/cleaninginspectoR","hunzikp/rtree", "impact-initiatives/koboloops",
                  "tidyr","dplyr", "ggplot2", "readr", "stringr", "lubridate", "readxl", "rgdal", "sf", "purrr", "sdcMicro")

## Running the packaging loading function
loadInstall_package(package_list)


source("01_PrepingAnalysis.R", encoding = "UTF-8")
source("utils.R")
source("analysisplan_factory.R", encoding = "UTF-8")
source("analysis_functions.R", encoding = "UTF-8")


## The commented lines below should be uncommented to run the script from scratch
cleaned_data_adm1$admin0 <- "BFA"
# analysisplan_admin_0 <- make_analysis_plan_template(df= cleaned_data_adm1,
#                                                     questionnaire = questionnaire,
#                                                     repeat.for.variable = "admin0",
#                                                     hypothesis.type = "direct_reporting",
#                                                     template_file = template_analysisplan_file
# )
# analysisplan_admin_0 <- analysisplan_admin_0[!is.na(analysisplan_admin_0$dependent.variable.type),]
# write_csv(analysisplan_admin_0, "data/analysis_plans/analysisplan_admin_0.csv")
# 
# 
# 
# final_result_admin_0 <- from_analysisplan_map_to_output(data = cleaned_data_adm1,
#                                                             analysisplan = analysisplan_admin_0,
#                                                             weighting = combined_weights_adm1,
#                                                             questionnaire = questionnaire)
# 
# saveRDS(final_result_admin_0, "outputs/final_result_admin_0.RDS")

cleaned_data_adm1$admin0 <- "BFA"

final_result_admin_0 <- readRDS("outputs/final_result_admin_0.RDS")

summary_stats_admin_0_final <- format_results(cleaned_data_adm1, final_result_admin_0, aggregate_level = "admin0", weights = "weights_sampling", csv = TRUE)
